requirements.txt
fastapi
uvicorn[standard]
face_recognition
faiss-cpu
numpy
opencv-python
